fix: sometimes input will cause crash
Guess: This is because the unmatched lifecycle of the closure connected to the signal and the worker object.
For example, when worker A finishes its job, A may be destroyed later, but the closure to handle `sendInput` signal may be left for a while. Then next time a `sendInput` signal still triggers the closure to execute its code, which makes it load an invalid address.

To solve it, we may not use the closure, but bind the signal to a slot of the worker. And the relative variables are set to be some members of the worker. As a result, when the worker is destroyed, no slot will receive `sendInput` signal again, and no invalid address is touched.